requirements.txt
aniso8601==3.0.2
backcall==0.1.0
boto3==1.9.32
botocore==1.12.32
decorator==4.3.0
Django==2.1.2
django-graphql-jwt==0.1.14
docutils==0.14
graphene==2.1.3
graphene-django==2.2.0
graphql-core==2.1
graphql-relay==0.4.5
ipython==7.0.1
ipython-genutils==0.2.0
jedi==0.13.1
jmespath==0.9.3
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
promise==2.2.1
prompt-toolkit==2.0.6
ptyprocess==0.6.0
Pygments==2.2.0
PyJWT==1.6.4
python-dateutil==2.7.4
pytz==2018.6
Rx==1.6.1
s3transfer==0.1.13
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
traitlets==4.3.2
urllib3==1.24
wcwidth==0.1.7
